# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Logic Circuit Simulator API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

/**
 * Circuit JSON data (gates, wires, positions)
 */
export type CircuitData = { [key: string]: unknown };

export interface Circuit {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** Circuit JSON data (gates, wires, positions) */
  data: CircuitData;
  /** @nullable */
  shareCode?: string | null;
  isPublic?: boolean;
  /** @nullable */
  tags?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CircuitInputData = { [key: string]: unknown };

export interface CircuitInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  data: CircuitInputData;
  tags?: string;
  isPublic?: boolean;
}

export type CircuitUpdateData = { [key: string]: unknown };

export interface CircuitUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  data?: CircuitUpdateData;
  tags?: string;
  isPublic?: boolean;
}

export interface ShareInfo {
  shareCode: string;
  shareUrl: string;
}

export interface CircuitStats {
  total: number;
  recentlyEdited: Circuit[];
  publicCount: number;
}