# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Logic Circuit Simulator API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: circuits
    description: Circuit operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /circuits:
    get:
      operationId: listCircuits
      tags: [circuits]
      summary: List all circuits
      responses:
        "200":
          description: List of circuits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Circuit"
    post:
      operationId: createCircuit
      tags: [circuits]
      summary: Create a new circuit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CircuitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Circuit"

  /circuits/{id}:
    get:
      operationId: getCircuit
      tags: [circuits]
      summary: Get a circuit by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Circuit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Circuit"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCircuit
      tags: [circuits]
      summary: Update a circuit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CircuitUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Circuit"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCircuit
      tags: [circuits]
      summary: Delete a circuit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /circuits/{id}/share:
    post:
      operationId: shareCircuit
      tags: [circuits]
      summary: Generate or regenerate a share link for a circuit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Share info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareInfo"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /circuits/shared/{shareCode}:
    get:
      operationId: getSharedCircuit
      tags: [circuits]
      summary: Get a circuit by its share code (read-only public access)
      parameters:
        - name: shareCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Shared circuit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Circuit"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /circuits/stats:
    get:
      operationId: getCircuitStats
      tags: [circuits]
      summary: Get dashboard stats
      responses:
        "200":
          description: Stats summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CircuitStats"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    Circuit:
      type: object
      required: [id, name, data, createdAt, updatedAt]
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: ["string", "null"]
        data:
          type: object
          description: Circuit JSON data (gates, wires, positions)
          additionalProperties: true
        shareCode:
          type: ["string", "null"]
        isPublic:
          type: boolean
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CircuitInput:
      type: object
      required: [name, data]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        data:
          type: object
          additionalProperties: true
        tags:
          type: string
        isPublic:
          type: boolean

    CircuitUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        data:
          type: object
          additionalProperties: true
        tags:
          type: string
        isPublic:
          type: boolean

    ShareInfo:
      type: object
      required: [shareCode, shareUrl]
      properties:
        shareCode:
          type: string
        shareUrl:
          type: string

    CircuitStats:
      type: object
      required: [total, recentlyEdited, publicCount]
      properties:
        total:
          type: number
        recentlyEdited:
          type: array
          items:
            $ref: "#/components/schemas/Circuit"
        publicCount:
          type: number