# lib/api-zod/src/generated/types/circuit.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Logic Circuit Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { CircuitData } from './circuitData';

export interface Circuit {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** Circuit JSON data (gates, wires, positions) */
  data: CircuitData;
  /** @nullable */
  shareCode?: string | null;
  isPublic?: boolean;
  /** @nullable */
  tags?: string | null;
  createdAt: Date;
  updatedAt: Date;
}