# lib/api-zod/src/generated/types/circuitInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Logic Circuit Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { CircuitInputData } from './circuitInputData';

export interface CircuitInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  data: CircuitInputData;
  tags?: string;
  isPublic?: boolean;
}