# lib/api-zod/src/generated/types/circuitStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Logic Circuit Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { Circuit } from './circuit';

export interface CircuitStats {
  total: number;
  recentlyEdited: Circuit[];
  publicCount: number;
}