# lib/api-zod/src/generated/types/circuitUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Logic Circuit Simulator API
 * OpenAPI spec version: 0.1.0
 */
import type { CircuitUpdateData } from './circuitUpdateData';

export interface CircuitUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  data?: CircuitUpdateData;
  tags?: string;
  isPublic?: boolean;
}