# lib/api-zod/src/generated/types/shareInfo.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Logic Circuit Simulator API
 * OpenAPI spec version: 0.1.0
 */

export interface ShareInfo {
  shareCode: string;
  shareUrl: string;
}